feat:CRUD operation using sweet Alert

+++ src/app/dashboard/product/list-product/list-product.component.ts
@@ -3,6 +3,9 @@ import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { AddEditProductComponent } from '../add-edit-product/add-edit-product.component';
 import { Product } from '../product';
 import { ProductService } from '../product.service';
+import { filter } from 'rxjs/operators';
+import Swal from 'sweetalert2';
+import { ExcelService } from '../excel.service';
 
 @Component({
   selector: 'app-list-product',
@@ -12,44 +15,84 @@ import { ProductService } from '../product.service';
 export class ListProductComponent implements OnInit {
   product!:Product[];
   constructor(public dialog:MatDialog,
-              private service:ProductService) {
+              private service:ProductService, public excelService:ExcelService) {
                 this.getProductsList();
                }
 
   ngOnInit(): void {
   }
 
+  //Adding new product in json file
   public addProduct(): void {
-    const dialogConfig = new MatDialogConfig();
-    dialogConfig.disableClose = true;
-    dialogConfig.autoFocus = true;
-    dialogConfig.width = "35%";
-    this.dialog.open(AddEditProductComponent, dialogConfig);
+    
+    let dialogRef = this.dialog.open(AddEditProductComponent);
+    
+    dialogRef.afterClosed().pipe(
+      filter(result => result )
+    ).subscribe(prod_name => {
+      this.product.push(prod_name);
+    })
   }
 
+  //Get all product list from the json file
   public getProductsList(){
     return this.service.getProducts().subscribe((data:Product[])=>{
       this.product = data;
     })
   }
 
-  //Get Employee by Id from Rest API
+  //Get product by Id from service
   public editProd(product: Product) {
-    this.service.getProductById(product.id).subscribe((data: Product) => {
-
+    debugger
       const dialogConfig = new MatDialogConfig();
       dialogConfig.disableClose = true;
       dialogConfig.data = product;
       dialogConfig.autoFocus = true;
       dialogConfig.width = "35%";
-      this.dialog.open(AddEditProductComponent, dialogConfig);   
-    })
+      let dialogRef = this.dialog.open(AddEditProductComponent, dialogConfig)
+      
+      dialogRef.afterClosed().subscribe(result => {
+        var check = this.product.findIndex(x => x.id === result.id);
+        this.product[check] = result;
+      });
   }
 
+  //Delete product from the list
   public deleteProd(id:number){
-    this.service.deleteProduct(id).subscribe((data: number)=>{
-      alert("Product delete successfully");
+    Swal.fire({
+      title: 'Are you sure want to remove?',
+      text: 'You will not be able to recover this data!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+
+        this.service.deleteProduct(id).subscribe((data: number)=>{
+          Swal.fire(
+            'Deleted!',
+            'Your imaginary file has been deleted.',
+            'success'
+          )
+          this.product.pop();
+        })
+       
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
     })
+    
   }
 
+  //Export table data into excel file
+  public exportAsExcel():void{
+      this.excelService.exportAsExcelFile(this.product, 'Product_data');
+  }
+
+  
 }
